This PR adds an optimization to the JSON reader improving its performance for S3 reads.

## The problem
Previously, the JSON reader would share a single file handle globally. Because that means that a single socket is shared over all threads, the network requests would be sequential. This is especially bad for S3 where per-connection throttling exists. For S3 it is therefore always recommended to download with multiple connections concurrently for max performance. 

## Some benchmarks 
I ran a few benchmarks to demonstrate the problem: (note: values are noisy, each benchmark was run once)

|File          |Type    |Size   |Copy method                     |Destination|Timing|
|--------------|--------|-------|--------------------------------|-----------|------|
|lineitem sf0.1|JSON    |207.4 MB|`CREATE TABLE test AS FROM "<file>"` |S3         |**17.54s** |
|lineitem sf0.1|JSON    |207.4 MB|`CREATE TABLE test AS FROM "<file>"` |local file  |0.20s  |
|lineitem sf0.1|JSON    |207.4 MB|`aws s3 cp "<file> ."` |S3         |3.48s  |
|lineitem sf0.1|JSON    |207.4 MB|`curl "<file>" --output test.json` |S3         |19.00s |
|lineitem sf1  |PARQUET |258.3 MB|`CREATE TABLE test AS FROM "<file>"` |S3         |4.91s |
|lineitem sf1  |PARQUET |258.3 MB|`CREATE TABLE test AS FROM "<file>"` |local file  |0.26s  |
|lineitem sf1  |PARQUET |258.3 MB|`aws s3 cp "<file>" . `|S3         |4.10s  |
|lineitem sf1  |PARQUET |258.3 MB|`curl "<file>" --output test.json` |S3         |20.00s |

We can clearly see here that for `CREATE TABLE test AS FROM <json_file_on_s3>` we are way slower than expected. Note that downloading the file with curl is similarly slow. `aws cli` downloads however are very fast: these use multiple connections for a download. Also note that  our parquet download is very fast, being quite close to the time it takes the aws cli to download a file.

### New result
With this PR, the time for `CREATE TABLE test AS FROM "<file>"` is reduced to **7.2s**, wich is over **2x faster**.  This was with a network speed of ~900MBps. At lower speeds this difference will likely be smaller.

Essentially, this optimization will increase performance up to the point where `number of threads * s3 throttle speed = available bandwidth` after that point, bandwidth becomes the bottleneck, as it should be :)

Note that this is still over 2x under the theoretical max speed of 3.48s + 0.20s = 3.68s (aws cli download time + local file json parse) This is probably due to the fact that for the sample_run we do a pret=ty hefty read which is not done in parallel with the other downloads.

## Future work
- Push performance even closer to optimal?
- investigate if the same applies to the CSV reader